Refactor: User repository
