+++ demo/src/main/java/com/project/demo/logic/AuthenticationService.java
@@ -1,6 +1,7 @@
 package com.project.demo.logic;
 
 import com.project.demo.entity.User;
+import com.project.demo.repository.UserRepository;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.crypto.password.PasswordEncoder;
